% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bind}
\alias{bind}
\title{Bind parameters to a function.
 
\code{bind} returns a closure with given parameters bound to it.}
\usage{
bind(func, ...)
}
\arguments{
\item{func}{Function prototype object \strong{required}.}

\item{...}{Unnamed or named parameters to bind \strong{required}.}
}
\value{

}
\description{
Bind parameters to a function.
 
\code{bind} returns a closure with given parameters bound to it.
}
\examples{
lie <- bind(cat, 'FRAUD REPORT:')  # bind
lie('looking for investors')       # have fun
lie('quick returns guaranteed')

}

