% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{bounds-package}
\alias{bounds-package}
\title{Bounds for causal and sufficient-cause interactions}
\description{
bounds: Bounds for causal and sufficient-cause interactions
}
\details{
This package contains functions for estimating bounds of causal
interactions and sufficient-cause interactions. Standard errors are also
estimated via bootstrap.
}
\author{
Arvid Sjolander, Xia Shen
}
\references{
If you use the \code{bounds} procedure in your analysis, please cite the following work:

Sjolander A, Lee W, Kallberg H, Pawitan Y. (2014). Bounds on causal interactions for binary outcomes. \emph{Biometrics} \bold{70}(3), 500-505.

If you use the \code{bounds.suff} procedure in your analysis, please cite the following work:

Sjolander A, Lee W, Kallberg H, Pawitan Y. (2014). Bounds on sufficient-cause interactions. \emph{European Journal of Epidemiology} \bold{29}(11), 813-820.
}
\seealso{
\code{bounds}, \code{bounds.suff}
}
\keyword{package}

