% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\docType{package}
\name{bounds}
\alias{bounds}
\alias{bounds-package}
\title{bounds: From Bind To Bounds}
\description{
Partial application via closures.
}
\section{Exports}{

  \code{\link{bind}}
  \code{\link{isBound}}
}

\section{Repository}{

  \url{https://github.com/chiefBiiko/bounds}
}

