\name{ex1}
\docType{data}
\alias{ex1}
\title{Simulated data for the examples}
\description{
The \sQuote{ex1} data contains a simulated dataset for the examples, with one binary response variable,
two exposure variable, and two covariates.
}
\format{ The \sQuote{ex1} data set contains an R \code{data.frame}.
}
\keyword{datasets}