% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{isBound}
\alias{isBound}
\title{Check whether a function is bound to its enclosure}
\usage{
isBound(func)
}
\arguments{
\item{func}{Function to check \strong{required}.}
}
\value{
Logical indicating whether func is bound, a closure.
}
\description{
\code{isBound} scans its input function's frame checking if there are any 
names not bound within its own scope - free variables. It returns a logical value 
indicating whether the input function is bound, aka a closure.
}
\details{
\code{func} can be any function except primitives. \code{isBound(func)} 
translates to: Is \code{func} bound to any names in its enclosing environment?
}

